Implement routing system within the app using vue router
