+++ src/router/index.js
@@ -1,0 +1,34 @@
+import { createRouter, createWebHistory } from "vue-router";
+import Home from "../views/Home.vue";
+
+const routes = [
+	{
+		path: "/",
+		name: "Home",
+		component: Home,
+		props: true,
+		children: [],
+	},
+	{
+		path: "/search/:param",
+		name: "search",
+		props: true,
+		component: Home,
+	},
+	{
+		path: "/about",
+		name: "About",
+		// route level code-splitting
+		// this generates a separate chunk (about.[hash].js) for this route
+		// which is lazy-loaded when the route is visited.
+		component: () =>
+			import(/* webpackChunkName: "about" */ "../views/About.vue"),
+	},
+];
+
+const router = createRouter({
+	history: createWebHistory(process.env.BASE_URL),
+	routes,
+});
+
+export default router;
